+++ .github/pull_request_template.md
@@ -1,0 +1,14 @@
+**Describe what the PR does:**
+
+**Does this fix a specific issue?**
+
+Fixes https://github.com/bachya/pypollencom/issues/<ISSUE ID>
+  
+**Checklist:**
+
+- [ ] Confirm that one or more new tests is written for the new functionality.
+- [ ] Update `README.md` with any new documentation.
+- [ ] Run tests and ensure 100% code coverage: `make coverage` (after running `make init`)
+- [ ] Ensure you have no linting errors: `make lint` (after running `make init`)
+- [ ] Ensure you have no typed your code correctly: `make typing` (after running `make init`)
+- [ ] Add yourself to `AUTHORS.md`.
