+++ pylintrc
@@ -1,0 +1,2 @@
+[SIMILARITIES]
+min-similarity-lines=10
