+++ pypollencom/allergens.py
@@ -15,22 +15,26 @@ class Allergens:
     @raise_on_invalid_zip
     async def current(self) -> dict:
         """Get current allergy conditions."""
-        return await self._request('get', 'forecast/current/pollen')
+        return await self._request(
+            'get', 'https://www.pollen.com/api/forecast/current/pollen')
 
     @raise_on_invalid_zip
     async def extended(self) -> dict:
         """Get extended allergen info."""
-        return await self._request('get', 'forecast/extended/pollen')
+        return await self._request(
+            'get', 'https://www.pollen.com/api/forecast/extended/pollen')
 
     @raise_on_invalid_zip
     async def historic(self) -> dict:
         """Get historic allergen info."""
-        return await self._request('get', 'forecast/historic/pollen')
+        return await self._request(
+            'get', 'https://www.pollen.com/api/forecast/historic/pollen')
 
     async def outlook(self) -> dict:
         """Get allergen outlook."""
         try:
-            return await self._request('get', 'forecast/outlook')
+            return await self._request(
+                'get', 'https://www.pollen.com/api/forecast/outlook')
         except RequestError as err:
             if '404' in str(err):
                 raise InvalidZipError('No data returned for ZIP code')
