+++ tests/test_client.py
@@ -26,13 +26,14 @@ async def test_request_error(aresponses, event_loop):
         'www.pollen.com', '/api/bad_endpoint/{0}'.format(TEST_ZIP), 'get',
         aresponses.Response(text='', status=404))
     aresponses.add(
-        'www.pollen.com', '/api/forecast/outlook/{0}'.format(TEST_ZIP),
-        'get', aresponses.Response(text='', status=500))
+        'www.pollen.com', '/api/forecast/outlook/{0}'.format(TEST_ZIP), 'get',
+        aresponses.Response(text='', status=500))
 
     with pytest.raises(RequestError):
         async with aiohttp.ClientSession(loop=event_loop) as websession:
             client = Client(TEST_ZIP, websession)
-            await client.request('get', 'bad_endpoint')
+            await client._request(
+                'get', 'https://www.pollen.com/api/bad_endpoint')
 
     with pytest.raises(RequestError):
         async with aiohttp.ClientSession(loop=event_loop) as websession:
